+++ simulation/robot/encoder/encoder.h
@@ -12,12 +12,12 @@ typedef struct encoder
     uint8_t sigB;
     uint8_t lastSigA;
     uint8_t counter;
-    int lastAngle;
+    double lastAngle;
     pthread_mutex_t* encoderMutex;
     pthread_barrier_t* encoderBarrier;
 } encoderStruct;
 
-void updateEncoder(encoderStruct* enc, int angle);
+void updateEncoder(encoderStruct* enc, double angle);
 void updateSigB(encoderStruct* enc, uint8_t* sigB);
 void incrementEncoder(encoderStruct* enc);
 void decrementEncoder(encoderStruct* enc);
